Extract WindowState and related types into window_state.rs
Move InitializerEditor, FunctionViewer, WindowKind, and WindowState
from main.rs into a dedicated window_state module. Add handle_undo
and handle_redo as methods on WindowState, replacing the free-function
calls from main.rs. Re-export moved types from main.rs so all
existing imports continue to work unchanged.
